% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_codes.R
\name{average_trt_effect}
\alias{average_trt_effect}
\title{Calculate \emph{Average Treatment Effect (ATE)} from the output of \code{\link{cappmx_fit}}.}
\usage{
average_trt_effect(result, burnin = 200)
}
\arguments{
\item{result}{Output of \code{\link{cappmx_fit}}.}

\item{burnin}{Number of burnin samples to discard}
}
\value{
The estimated \emph{ATE} from the CAPPMx fit.
}
\description{
Calculate \emph{Average Treatment Effect (ATE)} from the output of \code{\link{cappmx_fit}}.
}
