\name{CAPPMx-package}
\alias{CAPPMx-package}
\alias{CAPPMx}
\docType{package}
\title{
\packageTitle{CAPPMx}
}
\description{
An implementation of the Bayesian Nonparametric Common Atoms Regression by \insertCite{chandra_GBM22;textual}{CAPPMx}. See also \insertCite{mueller_cam23}{CAPPMx} for a review on Bayesian literature on historical control in clinical trials.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CAPPMx}
\packageIndices{CAPPMx}
}
\author{
\packageAuthor{CAPPMx}

Maintainer: \packageMaintainer{CAPPMx}
}
\references{
\insertAllCited{}
}
\keyword{ CAPPMx }
\examples{
## See the package-vignette file.
}