% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_codes.R
\name{get_auc}
\alias{get_auc}
\title{Investigating Population Equivalence}
\usage{
get_auc(
  cat_cov_trt = NULL,
  cont_cov_trt = NULL,
  cat_cov_rwd = NULL,
  cont_cov_rwd = NULL,
  result,
  burnin = 200,
  cv_prop = 0.15
)
}
\arguments{
\item{cat_cov_trt}{The matrix of categorical variables in the Treatment arm.}

\item{cont_cov_trt}{The matrix of continuous variables in the Treatment arm.}

\item{cat_cov_rwd}{The matrix of categorical variables in the RWD.}

\item{cont_cov_rwd}{The matrix of continuous variables in the RWD.}

\item{result}{Output of \code{\link{cappmx_fit}} function.}

\item{burnin}{Number of burnin samples to discard from \code{result}.}

\item{cv_prop}{The value of \eqn{p} for a leave-\eqn{100\times p \%}-out 
cross-validation for computing out-of-sample classification accuracy.}
}
\value{
A vector of AUC values corresponding to the proposed IS and 
random resampling schemes, respectively.
}
\description{
@details
 A randomly resampled set of covariates is selected from RWD with weights being the \emph{Importance-Resampling (IS)} ones
 derived from the output of \code{\link{cappmx_fit}}.
 The resampled subset and the covariate population in the treatment arm are merged 
 into a single dataset and then we try to classify patients in the merged sample as originally RWD or single-arm treatment cohort.
 For classification, we use Bayesian Additive Regression Tree (BART) \insertCite{bart2010}{CAPPMx} and report  the 
 area under the receiver operating characteristic curve (AUC) of the classification accuracy.
 For comparison, we also subsample randomly (instead of using the IS weights) from RWD
 to create another synthetic control arm and report the AUC.
}
\references{
\insertAllCited{}
}
