% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imp.variables}
\alias{imp.variables}
\title{Covariate names in the \code{MDACC_reproduced} dataset}
\format{
A string vector with the covariate names in the \code{\link{MDACC_reproduced}} dataset
}
\usage{
imp.variables
}
\description{
Covariate names in the \code{MDACC_reproduced} dataset
}
\examples{
data(imp.variables)
imp.variables
}
\keyword{datasets}
